pythonfile copy/requirement.txt
fastapi
uvicorn
numpy
pillow
tensorflow-cpu
keras
python-multipart